Start to make some common logic more common
